Refactor Disassembler and DisassemblyReport
Disassembler used to be a class with a single `Disassemble` function
which used to fill a member string variable. This commit changes that
into a free function returning that string.

Furthermore the output and input of this Disassemble function was changed into
two structs which combine the parameters.

DisassemblyReport was changed to not rely on a instance of Disassembler
anymore (which does not exist anymore). It now takes a line_to_address
mapping which is one of the resulting fields of the Disassemble call.