Add CodeView integration test to OrbitQtTests
This adds a very basic integration test for the web parts to
OrbitQtTests. The test will render the web page into an offscreen
QGraphicsScene.

It is meant to check whether a web page loads inside of Orbit. The test
can test the QWebChannel communication and check if that works as
expected.

Currently the setup uses the print request signal, to tell the C++ part
of the test that everything went as intended on the browser side.

If the signal fails to appear the test will timeout after 5 secs and
fail.