Make ScopedStatus hold weak_ptr of Executor
ScopedStatus used to hold a raw pointer to MainThreadExecutor, which
means it's the creators responsibility to ensure that the
MainThreadExecutor instance stays alive as long as the ScopedStatus
instance is alive.

This doesn't work well with the use-case of ScopedStatus. It is meant to
be used in background threads. Most of the API is thread-safe for this
purpose.

This commit fixes the problem by making ScopedStatus keep a weak_ptr of
MainThreadExecutor.