Tie lifetime of tasks to MainThreadExecutor
MainThreadExecutor::Schedule used to tie the lifetime of the scheduled
actions to `QApplication::instance()`. That means it can happen that a
scheduled action is executed after OrbitMainWindow and OrbitApp have
been destructed. But almost all tasks mutate resource in OrbitMainWindow
or OrbitApp, which means they shouldn't be executed after destruction of
these two.

This change ties the lifetime of MainThreadExecutor actions to the
lifetime of the MainThreadExecutor object.

By destructing the MainThreadExecutor instance before OrbitMainWindow
and OrbitApp, it can be ensured that the actions won't operate on
dangling pointers.